+++ frontend/guest-app/src/components/Question/QuestionCard/LikeButton/UndoLikeModal.js
@@ -3,7 +3,7 @@ import Grid from "@material-ui/core/Grid";
 import Button from "@material-ui/core/Button";
 import Typography from "@material-ui/core/Typography";
 import {Box} from "@material-ui/core";
-import CommonModal from "../../CommonModal/CommonModal.js";
+import CommonModal from "../../../CommonModal/CommonModal.js";
 
 function CancelButton(props) {
 	return (
